# Optimize ResourceContainer/AssemblyLoadEventHandler methods, remove allocs