--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/AppModel/ResourceContainer.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/AppModel/ResourceContainer.cs
@@ -249,7 +249,7 @@ namespace MS.Internal.AppModel
             {
                 AssemblyName assemblyInfo = new AssemblyName(assembly.FullName);
 
-                string assemblyName = assemblyInfo.Name.ToLowerInvariant();
+                string assemblyName = assemblyInfo.Name;
                 string assemblyKey = string.Empty;
                 string key = assemblyName;
 
@@ -328,7 +328,7 @@ namespace MS.Internal.AppModel
             {
                 string key = assemblyName + assemblyVersion + assemblyKey;
 
-                s_registeredResourceManagers.TryGetValue(key.ToLowerInvariant(), out rmwResult);
+                s_registeredResourceManagers.TryGetValue(key, out rmwResult);
 
                 // first time. Add this to the hash table
                 if (rmwResult == null)
@@ -348,7 +348,7 @@ namespace MS.Internal.AppModel
                         rmwResult = new ResourceManagerWrapper(assembly);
                     }
 
-                    s_registeredResourceManagers[key.ToLowerInvariant()] = rmwResult;
+                    s_registeredResourceManagers[key] = rmwResult;
                 }
             }
 
@@ -384,7 +384,7 @@ namespace MS.Internal.AppModel
 
         #region Private Members
 
-        private static Dictionary<string, ResourceManagerWrapper> s_registeredResourceManagers = new();
+        private static Dictionary<string, ResourceManagerWrapper> s_registeredResourceManagers = new(StringComparer.OrdinalIgnoreCase);
         private static ResourceManagerWrapper _applicationResourceManagerWrapper = null;
         private static FileShare _fileShare = FileShare.Read;
         private static bool assemblyLoadhandlerAttached = false;