--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/AppModel/ResourceContainer.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/AppModel/ResourceContainer.cs
@@ -49,7 +49,7 @@ namespace MS.Internal.AppModel
             {
                 if (s_applicationResourceManagerWrapper == null)
                 {
-                    // load main excutable assembly
+                    // load main executable assembly
                     Assembly asmApplication = Application.ResourceAssembly;
 
                     if (asmApplication != null)