@JeremyKuhne 2024-10-29
@adamsitnik wanted to point this out to you as it is doing some assembly name parsing in a hot path for WPF.